Move token randomizers into tests
